Function parsell::character_ref

pub fn character_ref<F>(f: F) -> CharacterRef<F>

An uncommitted parser that reads one character by reference.

The parser character(f) reads one character ch from the input, if f(&ch) is true then it commits and the result is ch, otherwise it backtracks.

This does not require characters to be copyable.