Trait parsell::HasOutput

pub trait HasOutput<Ch, Str> {
    type Output;
}

A trait for parsers which can infer their output type from their input types.

Associated Types

type Output

The type of the data being produced by the parser.