Trait parsell::ToPersistant

pub trait ToPersistant {
    type Persistant;
    fn to_persistant(self) -> Self::Persistant;
    fn from_persistant(p: Self::Persistant) -> Self;
}

A trait for data which can be saved to and restored from long-lived state.

The canonical example of this trait is Cow<'a,T> which can be saved to and restored from `T::Owned'.

This trait is lot like ToOwned, the difference is that Cow<'a,T>::Owned is Cow<'a,T>, not T::Owned.

Associated Types

type Persistant

Required Methods

fn to_persistant(self) -> Self::Persistant

fn from_persistant(p: Self::Persistant) -> Self